python/requirements.txt
numpy
scipy
pika
pyserial
tkinter
requests